right repo

+++ mainwindow.h
@@ -27,7 +27,7 @@ private slots:
     void DisplayStat(StatServer);
     void DisplayConnectStatus(uint16_t status);
     void DisplayError(uint16_t error);
-    void DisplaySuccess(uint16_t typeMess);
+    void DisplaySuccess(uint8_t typeMess);
 
 
     void on_cb_request_currentIndexChanged(int index);
